Switched from CSV to HDF5 using Pandas.
ALthough the HDF5 files is much bigger than the CSV file, it is easier to
handle as a single file, rather than two separate CSV files (one for
tracking and one for segmentation). The CSV writer is still there, but
I may make it a single npy file or a zip file.