Finally fixed batchtrack.py. Refactor SORTracker in the process.
Instead of having two implementations of SORTracker with and without
QObject base, I made it a single core class that is used both in GUI
and batch tracker.
The communication with the GUI and Qt threading is now done by the
wrapper class QSORTracker in sortrackerwidget.py, which also now
contains the widget for SORT tracker, SORTrackWidget.