Fixed cutility import. Added new distance metric.
Due to relative import change so far cutility (the C version of the
utility functions) was not getting imported and all the functions were
being computed in pure Python. Fixed that.

To merge overlaps, area of intersection as a fraction of the smaller
bbox is a better measure when animals are not supposed to overlap
much (often body parts get detected as different object). Added this.