Update of capture program and refinement for cython.
The capture program was very slow with ~10 FPS.
Tried to improve that by adding Cython version of the capture function.
On my laptop there is little difference in FPS with Cython and pure Python.

Also updated cutility.pyx with more specific data types for functions.
Made corresponding fixes in functions in tracking tool that call these.
Specifically pair-wise distance computation worked for both double and int (sortracker and yolactwidget)
Not sure if this really improves performance. I see 29 fps with pure Python and 30.x with cython.